requirements.txt
asn1crypto==1.5.1
boto3==1.34.104
botocore==1.34.104
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
moto==5.0.7
packaging==24.0
pg8000==1.31.2
pluggy==1.5.0
pycparser==2.22
pytest==8.2.0
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
s3transfer==0.10.1
scramp==1.4.5
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
xmltodict==0.13.0
